search contexts: refactor namespace filtering (#21445)
This PR introduces a `namespaces` parameter to the `searchContexts` query. Instead of matching by a single namespace, we now allow matching by multiple namespaces and returning their union. This allows us to simplify the API as well as the client.

With this we can also request the set of search contexts that are immediately associated with the current user (public instance-level contexts, user contexts, and member org contexts). (See first attempt: https://github.com/sourcegraph/sourcegraph/pull/21279)