auth: Prompt and allow re-authenticating with GitHub (#21495)
This change will prompt a user who has been opted in to private code after already being part of the public code trial. In this case, we need to prompt them to re-authenticate with GitHub to request the additional repo scope.

This change only updates the code host page and does not yet add a global banner.