oobmigrations: Startup check (#20967)
Out of band migrations that are incomplete and deprecated _at or on the current app version_ will cause the instance to shut down with a fatal log message.

On upgrade, the resolution is to restart the instance with the previous version and debug the reason the migration has not finished (could be it just needs to be ran longer; might be an error preventing progress). See #21670 for related documentation updates.