search: send public field back to zoekt indexserver (#21689)
We want zoekt to start recording if a repository is private or
public. So we add public to the list of index options we return for a
repository.

Note: the repo table stores "private", but we chose to call the field
"public". This is done in case a mistake is made and the field is
unset. Then go defaults to false => privacy leak.