phabricator: Remove custom json unmarshalling (#21702)
The custom unmarshaller on the Repo type was removed. Instead, we marshal into
an apiRepo struct and convert that into a Repo struct. Having a custom
unmarshal on the Repo type meant that when it was marshalled into our database
it couldn't be unmarshalled again. This in turn broke our ability to build a
clone url from a stored Phabricator repo.