Don't fetch unknown stargazerCount field on old GHE versions (#21864)
GitHub enterprise prior to version 3.0 doesn't support this field. This is causing the query to fail and currently all repo syncs to github enterprise < 3.0 are falling back to sequential mode, I _think_ this fixes it. There is also a stargazers connection on the repository, which could be queried to make it work on these older versions.

Found this log in our integration test suite:

```
2021-06-07T21:38:50.841746670Z 21:38:50              repo-updater | t=2021-06-07T21:38:50+0000 lvl=warn msg="github sync: fetching in batches failed. falling back to sequential fetch" error="error in GraphQL response: Field 'stargazerCount' doesn't exist on type 'Repository'"
```