Improve run time of GraphQL generator
The graphql generator uses babel to parse the AST to very precisely find usages of graphql-tag like `gql\``. Since we always use `gql\``, we can make a simplification and not rely on the tag plucking. This reduces the generate time from 11s to 5.4s locally for me. Also, this is the same improvement that we already did for watch mode.