gilab: Backoff correctly when sync is forbidden
Currently, we're seeing quite a few user added GitLab code hosts on
Cloud that have never synced. It appears that most of the cases are
caused by GitLan returning a 403 Forbidden when attemping to list repos.
Our current code handles this as a hard failure and bails out without
writing to the last_sync column. This in turn means that the code host
sync will be attempted again in about a minute.

Instead, we handle 403 errors in the same way as Unauthorized errors and
treat it as if the code host returned a list of zero repos. This allows
us to "complete" the sync and means we'll properly back off with sync
attempts.