Make FuzzyModal storybook fully interactive
Previously, the storybook for FuzzyModal lost the focus on the input on
every keystroke because the storybook forced a refresh on every state
change for the query, focus index or the max results. Now, those
properties have been refactored into internal state of the FuzzyModal
component so that we no longer need to force a re-render in React.

The original motiviation to make `query`, `focusIndex` and `maxResult`
into properties was to make it easier to create a storybook. This commit
in essence reverts that decision because it turned out to be easier if
those variables are internal state instead.