Query input: select completion item on Enter keypress
Previously, pressing enter in the query editor would always trigger a
search including when the editor displayed completion suggestions. Now,
the query editor selects the completion suggestion instead.

With this change, it's possible to type `lang:T`, press enter to select
the `lang:TypeScript` completion and enter again to complete the search.
Previously, a Tab press was required to select the completion item.
The new behavior is more consistent with the defaults in VS Code and
IntelliJ.