scheduler: Always dequeue high priority repos for update.
Before, when disableAutoGitUpdates was set we would stop our update
scheulder. Now, we always run the scheduler but when
disableAutoGitUpdates is set we only handle repos flagged as high
priority. In practice, these are repos added from outside out scheduler,
specifically from the refresh webhook.

This allows customers to disable our scheduler but continue to manually
update repos when required.