codeintel/resolvers: Replace github.com/pkg/errors with cockroachdb
This appears in a string as part of a diff we use in tests. Adding the
CI check to check for github.com/pkg/errors only within the import()
statements is slightly more complicated and more involved that it is
probably worth. Changing this shouldn't cause any changes in behaviour
for us.