LSIF: protocol: new documentationResult identifier approach
Today, a URL for a Go symbol in API docs is e.g.:

```
/cmd-frontend-auth#typemocksCreateUserAndSave
```

But as a Go developer, I would rather have the URL be:

```
/cmd/frontend/auth#mocks.CreateUserAndSave
```

The package path `cmd-frontend-auth` -> `/cmd/frontend/auth` is easily solved in the lsif-go indexer.

The URL hash `typemocksCreateUserAndSave`, however, is the joined _slugs_ `["type", "mocks", "CreateUserAndSave"]` from each hierarchial `documentationResult`, meaning:

* We can't get rid of "type" (we have no idea what that is)
* We don't know if it's right to join slugs with a `.` (that is right for Go, but perhaps for C++ you need `::` or `->` or something contextual even.)

With this change, once we get to a page level the URL hash can effectively be chosen by the language indexer. It can emit `mocks.CreateUserAndSave` knowing that is a good unique identifier within that page and with the context of what those mean in Go.

Today, the code dealing with the `documentationResult` graph starting at the root of the project have a few awkward choices to make:

1. Should we respect the `slug` of the root documentation result?
1. What should the `slug` of the root documentation result be?
2. What should the `label` of the root documentation result be?
3. What should the content of the root documentation result be?

This adds guidelines for each of these questions, and reduces complexity in some ways (such as never needing to respect the `slug` of the root documentation result, because it is now always `""`.)

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>