internal/search: order results by repository priority (star count).
This makes result pages more stable and useful.

This commit takes the partially-ordered stream of results from the zoekt
backends and merges them using two new pieces of information added to
the API: the priority of the shard that a zoekt.SearchResult comes from,
and the maximum shard priority that the backend has not yet returned
results for.