search frontend: new experimental sign up page (non functional) (#22104)
Part of #21093

Basic, non-functional experimental sign up page that will be used to A/B test user sign up CTAs. 

Page is only shown if:
- The user is on Sourcegraph.com
- The `w0-signup-optimisation` feature flag is enabled for the user
- The user is not signed in
- The user navigates to the `/sign-up` page with a `src` query parameter that contains the source of the sign up (eg. `/sign-up?src=monitoring` for code monitoring CTA)

For now, links and buttons don't do anything. Functionality will be added in a later PR.

The feature flag does not exist yet so no user should see this page for now; I have added Storybook tests to see what the end result might look like. Note that the new customer logos images in the Storybook tests might not show right until this change is deployed to Sourcegraph.com since Storybook uses images on sourcegraph.com.

Because this is an experimental feature intended for short term A/B testing, some testing and cleanup has been skipped (eg. unit tests, extracting colors to variables, pixel perfect design implementation, etc). If we decide that this is a feature we want to keep, we can work on polishing it.