limit event matches before fetching repo metadata (#22144)
With the initial implementation, we'd batch events, then fetch metadata
for all results in the batch before limiting the batch. For queries with
very large result sets that return quickly, this means we're doing extra
work by fetching metadata for many results that won't actually be
returned. This can cause a significant delay. This commit changes it so
we limit first, then fetch metadata.