Move line range from hash to query parameter (#22058)
To support OpenGraph preview images we need access to the blob line range on the backend. Currently, line range is store in the hash part of the url, which is not set to the server. In this PR, I moved the line range from hash to a value-less query parameter, e.g: `?L1:3`.