search: fix overLimit for type:repo search (#22094)
We set overLimit to true whenever we len(repos)>=limit. However, for a
global type:repo query we always have len(repos) = limit+1, which is why
we always show an alert.

Here we remove the alert from defaultRepositories and adjust the
condition at the same time. This means we will never show an alert for
type:repo search. We should think about whether we should still send a
signal to the user that the results are possibly incomplete (aka '+'
sign)

The best thing would be to follow an earlier proposal by Rjinard to map
type:repo to repo: queries.

Co-authored-by: Tomás Senart <tomas@sourcegraph.com>