LSIF: conversion: allow correlation without pruning
I am working on a better, more extensive test suite for the Sourcegraph
documentation extension to LSIF. Specifically, I am making a better
version of what is in the lsif-go indexer currently (generating static
markdown files from the LSIF data for snapshot testing.)

I would like to use `lib/codeintel/lsif/conversion` directly because it
does a lot of the annoying data correlation for me. But in the context
of snapshot testing, data pruning is not relevant (and could hide useful
information) so this commit makes it optional.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>