API docs: LSIF: conversion: reject duplicate pathIDs
Prior to this change, duplicate pathIDs for a given bundle would result
in no error from the conversion process itself, but would fail to insert
to the DB due to unique constraints imposed there.

This adds logs that inform an indexer developer of which keys are duplicates
and why the API docs in their bundle are being rejected. It also enables the
[lsif-static-doc](github.com/sourcegraph/lsif-static-doc) tool to not emit
documentation and log errors in this case to make it abundantly clear that
the API docs in the bundle are invalid.

This doesn't affect the bundle overall, it just means API docs in the bundle
will be rejected outright.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>