worker: Prevent panic if database initialization fails
Only perform a type asserion if the error was `nil`.

A panic occurs if the database could not be initialzed, and the cause of
the error is not provided.

This is because an unsafe type assertion is performed regardless of what
is returned by `initDatabaseMemo`.

See https://github.com/sourcegraph/sourcegraph/issues/22286.