return `[]result.Match` from search functions (#22362)
* return []result.Match from searchSymbolsInRepo

This changes searchSymbolsInRepo to return []result.Match instead of
[]*result.FileMatch. This is less specific, but removes the need for
conversions down the line.

* return []result.Match for textsearch

This returns []result.Match instead of []*result.FileMatch from
textsearch methods. This removes the need for a conversion down the
line, and allows us to get rid of the conversion helper
fileMatchesToMatches at the expense of slightly less specific return
values.