API docs: LSIF: collect and store information about documentation page pathIDs
Today, the GraphQL API and LSIF DB schema for API docs expose and store combined
information describing a hierarchy of documentation pages in a repository. That is,
you can request a page a specific path and get its exhaustive information (including
its children.)

But navigating the full structure of the documentation pages in a repository would
be quite a slow operation - you'd need to pull every documentation page out of the
repository / DB.

This change makes us store a separate set of path information only - similar to a
directory structure but with slightly more info - about the documentation pages in
a repository. This makes it possible for us to e.g.:

* Have an documentation overview page that shows all (or N levels deep) of documentation
  pages in a repository, rather than just the pages immediately below the root (which
  are sometimes just index pages and worth omitting.)
* Use this data for better tree sidebar navigation.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>