Fix race condition in log output by writing to same buffer
Before this change the logging code worked "by accident": some writes
went to `os.Stdout` (through `fmt.Print`) and others went to a buffer
and *then* to `os.Stdout`.

But these writes were not ordered, because they depend on how often
`os.Stdout` is flushed.

That lead to `sg` sometimes printing logs out of order: the end of the
line before the start of the line.

This fixes that.