dbtest: Always re-create template database (#22419)
I spent hours debugging an issue where the latest migration I was
working on didn't get applied properly. The issue was that the template
database was being re-used, and it was previously created off of my
local default database, so the schema_migrations version was already the
latest. That meant no migrations were applied, even though the content
of the migrations had changed.