feat(sg): Improve 'sg migration fixup'
`sg migration fixup` now properly supports updating the database on your
local machine, as well as fixing any conflicting files.

Many improvements were made to the way we were running other `sg
migration` commands, including moving `up` and `down` to use the actual
`golang-migrate` library, as opposed to the executable (which required
an additional installation of a tool).

Finally there was some larger refactoring of the `sg` tool, as it was
growing too big for a single `package main`.

Note: I split the `migration.go` file into two different packages, the
first being `package migration`, which focuses ONLY on the migrations
and `package squash` which is used to run the docker container that we
use for squashing a bunch of migrations into one migration (which is a
completely different task).