API docs: LSIF: prevent `null` lists in GraphQL responses
Prior to this change, children and tag arrays could sometimes be `null` in our GraphQL
JSON response data. This was sometimes incorrect (in the case of tags), and other times
just annoying to deal with. This change makes us do our best to prevent `null` lists
from appearing in our data, preventing a myriad of bugs.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>