OpenGraph preview: Add function to retrieve a symbol at specific line and character (#22291)
Adds a function which returns a single symbol (if exists) in a file at a specific line and character number. It iterates all available symbols in the file and returns the first symbol with a matching range.

The function to retrieve a symbol is also exposed through the GQL API. It is added as an additional resource on the File type.