batches: add and use a new field for the desired publication state (#21092)
* batches: make the published field optional

For now, we'll just treat a missing field as being equivalent to being
unpublished.

Part two of #18277.

* Fix formatting.

* database: add migration for desired publication state

* WIP of looking at desired state

TODO: collapse closing into this.

* Updates to plan based on the UI state.

* Revert unneeded change.

* Correctly handle missing published fields in changeset specs.

* Output published values correctly in GraphQL.

* Typo fix.

* Add published == nil resolver test.

* Add publicationStateCalculator test.

* migrations: update previous changeset field modifications

This is a fun one: this PR needs to add a new column to `changesets`
that is also visible from the `reconciler_changesets` view. That view is
essentially a `SELECT * FROM changesets WHERE ...`. PostgreSQL uses the
columns that are present when the view is created to define the columns
available in the view.

The problem here is that, in recreating the view, we now include columns
that didn't exist the last time the view was recreated that were added
by previous migrations. When the migration test attempts to run and work
its way down the migrations, those migrations now fail when they try to
drop the column, since it is still referred to by
`reconciler_changesets`.

The fix here is to recreate the view each time `changesets` has a field
added or removed. Changing the old migrations prevents errors if
Sourcegraph is migrated up to version 1528395838 (the version in this
PR) and then down-migrated again back past version 1528395798.