dbconn: Better SQL pool stats
This commit makes use of the sqlstats package to extend the set of
metrics we capture related to the SQL connection pool. In particular,
we're going to track down if there's significant wait time for a free
connection regularly, which should indicate we need to increase the
maximum number of open connections.

Additionally, we now correctly specify application name to the database.