respect passed-in api limits in searcher
This commit modifies searcher to remove hard-coded file match limits and
line match limits in favor of using the limit passed in through the API.

On performance, I tested many different ways of limiting matches in the
parallel hot loop, and this performed by far the best. I did this inline
rather than making a nice named function because mutating the file match
by passing a pointer to it was causing the file match to escape the
stack, creating a huge spike in allocations.