dbconn: Close connections by idle time (#22586)
This commit replaces our usage of SetConnMaxLifetime with
SetConnMaxIdleTime which was later added to the standard library and
beter captures our intent when we introduced the former.

This might also improve the connection wait time being reported. The
theory is that because connections get closed lazily with
SetConnMaxLifetime, some connection requests can wait longer when an
underlying connection expires.

Only closing idle connections should address this.