users: Disallow new users with existing verified email (#22613)
We should not allow new users to sign up with an email address that is
already verified and assigned to another user.

We chose not to add a new constraint as it would require a migration
and it's not obvious what to do with existing accounts that would now
be considered invalid. It's simpler to just not allow this moving forward.