Initial drilling of UI publication state through the UI.
This doesn't yet do anything actually useful: it just ensures we have
the data we need at the points we need it to determine which UI to show.

The major unanswered question right now: how do we get the checkbox
state from the list elements to the alert that handles the backend call?
My current inclination is probably to move the state management to
PreviewList, then have _it_ invoke the mutation instead of the alert.