add file:contains.content predicate (#22666)
This adds a file predicate to filter to only files that contain content
matching the given pattern.

In textsearch cases, it's usually simpler to just use an "and" operator.
For example, `file:contains.content(q1) q2` is equivalent to `q1 and
q1`.

However, for non-file results, that's not possible. Consider "find diffs
that modify files that contain q1". With this change, that query can be
written as `type:diff file:contains.content(q1)` whereas that would
previously be unrepresentible in our query language.