indexed search: refactor zoektSearch
This is a non-mechanical refactor of indexed_search.zoektSearch to
make upcoming changes to zoektSearch easier.

Currenty zoektSearch handles 4 cases:
1. scoped search w/ select:repo
2. scoped search w/o select:repo
3. global search w/ select:repo
4. global search w/o select:repo

With this refactor we split zoektSearch into 2 separate functions, one
for scoped search and one for global search, zoektSearchGlobal.

In a next PR, we will rewrite zoektSearchGlobal completely to leverage
visibility queries.