extsvc/github: Fix isArchived and isLocked field fetching
I noticed this when working on streaming repo syncing. Because of the
JSON tags we set in those fields in order to avoid updating test
fixtures, we ended up not unmarshalling those fields from actual GraphQL
API requests.

REST API requests were not affected because we convert a restRepository
to a Repository.