security: Insert events for role change attempts
This commit adds two new security event types:
- RoleChangeDenied
- RoleChangeGranted

At the moment, if a user sends an API request to change a user's role
to admin for example, while they are **already** an admin, we will
still end up recording an event for RoleChangeGranted (provided that
all checks pass and the write to the database is successful). This is
because we make no initial checks for current state before writing the
new expected state to the database. The fact that we want to add this
check or not remains outside the scope of this commit and subject to
further deliberation.