Configure a default failure handler for GitHub OAuth.
Currently we don't have a failure handler registered for GitHub OAuth.
That results in customers being dropped off at a plaintext page for
actions like canceling their OAuth integration midway[0].

This change fixes a number of paths, but there are others that can lead
to the dreaded plaintext that are out of scope for this immediate change.

[0] https://sourcegraph.atlassian.net/browse/COREAPP-25