search: return 40 repo filters and 40 other
Previously we only returned 12. This was a hueristic value chosen based
on the old search UX which used horizontal space to represent these
values. We now have the sidebar (vertical space), so can display many
more.

Our implementation for computing these values is NlogK. We are changing
K from 12 to 40, so should not have a noticeable affect on performance.