executor: pass HOME/PATH to src-cli if Firecracker is disabled
This implements the solution discussed in [this comment][comment].

What problem this solves: when the `executor` is executed on a developer
machine then Firecracker is very likely disabled (and can't even be
enabled on macOS). When that's the case the `CliSteps` of an executor
job are executed on the host machine directly.

But that won't work with `src batch [...]` commands since those need
`HOME` and `PATH` to be set, which they are not, because when executing
on the host the environment of `src` consists only of the env vars
set in the `CliStep.Env` field. (That's how Go's `exec.Command.Env`
field works: once values are set, nothing else is inherited from the
parent process).

What this change here does is to fix that special case by checking
whether Firecracker is enabled and if not setting the env vars.

[comment]: https://github.com/sourcegraph/sourcegraph/pull/22726#discussion_r667766891