repos: Make streaming syncer the default
This commit makes the streaming syncer the default mode, while giving
admins the option to disable it by setting the env var to false.

We've been running this in production for about a week without any
major problems apart from a small bug that only affected sourcegraph.com
in practice (that was fixed in #22947).

Making it the default will benefit large customers already and allow us
to start deleting old code paths.