add EventPathMatch
This adds EventPathMatch as a type to our streaming API. This was
previously represented as an EventFileMatch with no line matches. The
intent is to better represent our result domain. Additinally, this
allows us to add metadata to path matches that isn't needed for content
matches.

This just makes the change for streaming events, but I will also
propagate this change down to our internal/search/result types.