rename File to Content in stream events
This commit renames our stream events to use the term "content" in the
places that it would use "file". This does not yet modify our result
types with the same convention, but that will come as a followup change.
The intent is to make the different events better match up with our
result types, so now we have Content, Symbol, and Path events that used
to be represented by the catch-all File event.