Multiplex queue handlers
Having one queue handler per queue makes it so we need to store less state along with a job, the queueName will always be correct. This also moves heartbeats to be a per-queue thing. This is not exactly a step forward but we didn't support multi-queue attachments anyways, so we can revisit once we want to do that.