Cleanup executor queue and move state to DB (#23165)
This PR updates the executor queue to do not hold state in-memory and removes now redundant cleanup logic. We don't need to manually keep track of records anymore, as we now have proper heartbeats coming from the executor itself. So we could get rid of most of the logic and just rely on the resetter now. Additionally, we now have one handler per queue now which makes it easier to track records as their ID is unique now (wasn't across queues before). That way, using the executorName as the worker_hostname, we can get all the relevant information from the DB and don't need to hold the state in-memory. This also extends the dbworker.Store by additional options to validate the worker is still the correct owner of the record (we did that before checking the in-memory map).