search: split global indexed search in public/private queries (#23118)
This changes how we query Zoekt for global searches. Instead of querying
the entire universe of indexed repos with 1 query and filtering the
results based on resolved repos, we send 2 queries: (1) a query
targeting all public repos (2) a query targeting all user accessible
private repos.

This has following advantages

1. We can skip repo resolution and filtering for query (1).
2. Query (2) will usually target a small subset of all repositories
which makes querying and filtering fast.

However, without repo resolution we don't validate the results coming
back from Zoekt, which could potentially lead to issues if the IDs and
repo names in Zoekt are out of sync with Sourcegraph. To guard against
this, we drop results if we fail to match them against repositories
during metadata hydration.