Continuously update executor log entries instead of writing at end (#23126)
This adds an `UpdateExecutionLogEntry` method to the `workerutil.Store` interface and implements it in `dbworker.Store` and in the executor `apiclient`.

The single user of this method so far is `command.Logger`. It continuously updates the output produced by a command in the database by calling `UpdateExecutionLogEntry` in a syncer goroutine that receceives the command output as it comes in and then flushes it to disk by calling `UpdateExecutionLogEntry`.