gitserver: use os.Remove in testRepoCorrupter
This way we handle an error if the file is already gone. Noticed it
because we are not setting PATH in cmd test wrapper, so depending on
your environment it might not find "rm". An example is nix where
rm (coreutils) is in a separate bin to sh (bash). We could just specify
PATH, but I thought it's better to just use os.Remove anyways.