dbtest: fallback to PGDATASOURCE if specified
We use PGDATASOURCE as a single environment variable for the DSN in
several places. Our dbtests did not respect it though, instead they
would try and mux in PG* variables into a hardcoded string.

Motivation: I prefer having a single envvar. Additionally it allows me
to do things like use a unix socket instead of TCP for a local
postgres. This is a path I am exploring for simplifying postgres in our
dev environments.